+++ validation/govalidator/validator.go
@@ -1,17 +1,12 @@
-package validation
+package govalidator
 
 import (
-	"reflect"
-	"strings"
-
 	"github.com/go-playground/locales/en"
 	ut "github.com/go-playground/universal-translator"
 	"github.com/go-playground/validator/v10"
 	en_translations "github.com/go-playground/validator/v10/translations/en"
+	"github.com/go-seidon/provider/validation"
 )
-
-type CustomValidator = func(fl validator.FieldLevel) bool
-type CustomTagName = func(field reflect.StructField) string
 
 type goValidator struct {
 	client     *validator.Validate
@@ -27,23 +22,23 @@ func (v *goValidator) Validate(i interface{}) error {
 
 	errs, ok := err.(validator.ValidationErrors)
 	if !ok {
-		return &ValidationError{
+		return &validation.ValidationError{
 			Message: err.Error(),
 		}
 	}
 
 	if len(errs) == 0 {
-		return &ValidationError{
+		return &validation.ValidationError{
 			Message: err.Error(),
 		}
 	}
 
-	return &ValidationError{
+	return &validation.ValidationError{
 		Message: errs[0].Translate(v.translator),
 	}
 }
 
-func NewGoValidator() *goValidator {
+func NewValidator() *goValidator {
 	en := en.New()
 	uniTrans := ut.New(en, en)
 	translator, _ := uniTrans.GetTranslator("en")
@@ -51,20 +46,10 @@ func NewGoValidator() *goValidator {
 	client := validator.New()
 
 	en_translations.RegisterDefaultTranslations(client, translator)
-	client.RegisterTagNameFunc(LabelTagName())
+	client.RegisterTagNameFunc(FieldTagName("label"))
 
 	return &goValidator{
 		client:     client,
 		translator: translator,
-	}
-}
-
-func LabelTagName() CustomTagName {
-	return func(field reflect.StructField) string {
-		name := strings.SplitN(field.Tag.Get("label"), ",", 2)[0]
-		if name == "-" || name == "" {
-			return field.Name
-		}
-		return name
 	}
 }
